+++ podchat/src/main/java/com/fanap/podchat/chat/Chat.java
@@ -1167,6 +1167,12 @@ public class Chat extends AsyncAdapter {
             case Constants.TURN_ON_VIDEO_CALL:
                 handOnCallParticipantAddedVideo(chatMessage);
                 break;
+            case Constants.START_SHARE_SCREEN:
+                handOnShareScreenStarted(chatMessage);
+                break;
+            case Constants.END_SHARE_SCREEN:
+                handOnShareScreenEnded(chatMessage);
+                break;
             case Constants.TURN_OFF_VIDEO_CALL:
                 handleOnCallParticipantRemovedVideo(chatMessage);
             case Constants.GET_CALLS:
@@ -2218,6 +2224,38 @@ public class Chat extends AsyncAdapter {
         return uniqueId;
     }
 
+    //Send request to start screen share screen
+    public String startShareScreen(long callId) {
+
+        //TODO start share screen streaming
+        String uniqueId = generateUniqueId();
+        if (chatReady) {
+            String message = CallAsyncRequestsManager.createStartShareScreenMessage(callId, uniqueId);
+            setCallBacks(false, false, false, true, Constants.START_SHARE_SCREEN, null, uniqueId);
+            sendAsyncMessage(message, AsyncAckType.Constants.WITHOUT_ACK, "SEND_START_SHARE_SCREEN");
+        } else {
+            onChatNotReady(uniqueId);
+        }
+
+        return uniqueId;
+    }
+
+    //Send request to end screen share screen
+    public String endShareScreen(long callId) {
+
+        //TODO stop share screen streaming
+        String uniqueId = generateUniqueId();
+        if (chatReady) {
+            String message = CallAsyncRequestsManager.createEndShareScreenMessage(callId, uniqueId);
+            setCallBacks(false, false, false, true, Constants.END_SHARE_SCREEN, null, uniqueId);
+            sendAsyncMessage(message, AsyncAckType.Constants.WITHOUT_ACK, "SEND_END_SHARE_SCREEN");
+        } else {
+            onChatNotReady(uniqueId);
+        }
+
+        return uniqueId;
+    }
+
     // TODO: 1/31/2021 Create new view and send with call back
     private boolean hasRemotePartnerView() {
 
@@ -2298,6 +2336,42 @@ public class Chat extends AsyncAdapter {
             addVideoCallPartner(response, true);
 
         listenerManager.callOnCallParticipantStartedVideo(response);
+
+    }
+
+    private void handOnShareScreenStarted(ChatMessage chatMessage) {
+
+        if (sentryResponseLog) {
+            showLog("RECEIVE_SHARE_SCREEN_STARTED", gson.toJson(chatMessage));
+        } else {
+            showLog("RECEIVE_SHARE_SCREEN_STARTED");
+        }
+
+        ChatResponse<JoinCallParticipantResult> response = CallAsyncRequestsManager.handleOnParticipantJoined(chatMessage);
+
+        if (podVideoCall != null) {
+
+        }
+
+        listenerManager.callOnShareScreenStarted(response);
+
+    }
+
+    private void handOnShareScreenEnded(ChatMessage chatMessage) {
+
+        if (sentryResponseLog) {
+            showLog("RECEIVE_SHARE_SCREEN_ENDED", gson.toJson(chatMessage));
+        } else {
+            showLog("RECEIVE_SHARE_SCREEN_ENDED");
+        }
+
+        ChatResponse<JoinCallParticipantResult> response = CallAsyncRequestsManager.handleOnParticipantJoined(chatMessage);
+
+        if (podVideoCall != null) {
+
+        }
+
+        listenerManager.callOnShareScreenEnded(response);
 
     }